feat: First Class Diagram Iteration
