+++ entregas/javierCader/examenFinal/factories/index.py
@@ -1,0 +1,7 @@
+
+from models.index import Patient
+
+class PatientFactory:
+    @staticmethod
+    def create_patient(name, discharge_date):
+        return Patient(name, discharge_date)
