--- entregas/javierCader/reto-007/docs/class_diagram.puml
@@ -1,33 +1,0 @@
-@startuml Class_Diagram
-
-class Paciente {
-    - nombre: String
-    - fechaAlta: Date
-    - encuestas: Encuesta
-}
-
-class Encuesta {
-    - fechaToma: Date
-    - dias: Día[]
-}
-
-class Día {
-    - numero: Int
-    - ingestas: Ingesta[]
-}
-
-class Ingesta {
-    - tipo: String
-    - alimentos: Alimento[]
-}
-
-class Alimento {
-    - nombre: String
-}
-
-Paciente "1" -- "0..*" Encuesta
-Encuesta "1" -- "1..5" Día
-Día "1" -- "0..*" Ingesta
-Ingesta "1" -- "0..*" Alimento
-
-@enduml
